Move P11Config to p11engine.h
And make p11engine.h usable without BUILD_P11=on
