+++ src/keymanager.h
@@ -6,19 +6,17 @@
 
 #include "httpinterface.h"
 #include "invstorage.h"
-#ifdef BUILD_P11
 #include "p11engine.h"
-#endif
 #include "types.h"
 #include "utils.h"
 
 // bundle some parts of the main config together
 // Should be derived by calling Config::keymanagerConfig()
 struct KeyManagerConfig {
+  P11Config p11;
   CryptoSource tls_ca_source;
   CryptoSource tls_pkey_source;
   CryptoSource tls_cert_source;
-  std::string p11_uptane_key_id;
   KeyType uptane_key_type;
   CryptoSource uptane_key_source;
 };
